+++ exts/vscode-tailwind-preview/README.md
@@ -1,0 +1,1 @@
+# vscode-tailwind-preview
