--- exts/webview-sample/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-  "editor.insertSpaces": false
-}
