--- exts/webview-sample/media/jsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2020",
-    "jsx": "preserve",
-    "checkJs": true,
-    "strict": true,
-    "strictFunctionTypes": true,
-    "lib": ["dom"]
-  },
-  "exclude": ["node_modules", "**/node_modules/*"],
-  "typeAcquisition": {
-    "include": ["@types/vscode-webview"]
-  }
-}
