--- exts/webview-sample/media/reset.css
@@ -1,30 +1,0 @@
-html {
-  box-sizing: border-box;
-  font-size: 13px;
-}
-
-*,
-*:before,
-*:after {
-  box-sizing: inherit;
-}
-
-body,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-ol,
-ul {
-  margin: 0;
-  padding: 0;
-  font-weight: normal;
-}
-
-img {
-  max-width: 100%;
-  height: auto;
-}
