--- exts/webview-sample/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2020",
-    "lib": ["es2020"],
-    "outDir": "out",
-    "sourceMap": true,
-    "strict": true,
-    "rootDir": "src"
-  },
-  "exclude": ["node_modules", ".vscode-test"]
-}
